requirements.txt
appdirs==1.4.3
atomicwrites==1.4.0
attrs==19.3.0
backports.functools-lru-cache==1.6.1
bcrypt==3.1.7
bleach==3.1.5
blinker==1.4
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cheroot==6.5.4
cherry==0.5.6
CherryPy==18.5.0
click==6.7
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
docopt==0.6.2
docutils==0.16
dominate==2.3.1
env-file==2020.5.9
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.3.0
Flask-JinjaHelpers==0.3.8
Flask-Login==0.5.0
Flask-LoginManager==1.1.6
Flask-Mail==0.9.1
Flask-Scrypt==0.1.3.6
Flask-Session==0.3.1
flask-session-cookie-manager==1.2.1
Flask-SQLAlchemy==2.4.1
Flask-Toastr==0.5.4
Flask-WTF==0.14.3
FormEncode==2.0.0a1
Genshi==0.7.3
gitignore==0.0.8
guess-language==0.2
httplib2==0.7.4
idna==2.8
importlib-metadata==1.6.1
importlib-resources==2.0.0
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
jieba==0.42.1
Jinja2==3.0.0a1
joblib==0.15.1
kiwisolver==1.2.0
Mako==0.7.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mdown==2019.4.13
mkdir==2019.4.13
more-itertools==8.3.0
numpy==1.18.1
orderdict==2019.9.25
packaging==16.8
pandas==1.0.0
pandas-schema==0.3.5
passlib==1.7.2
Paste==3.2.2
pkcs7==0.1.2
pluggy==0.13.1
portend==2.6
public==2019.4.13
py==1.8.1
pycparser==2.20
pycryptodome==3.9.6
Pygments==2.6.1
pyodbc==4.0.27
PyODBCSQLServer2000Database==0.2.1
pyparsing==2.2.0
pypiwin32==223
pytest==5.4.3
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
readme==0.7.1
readme-md==2019.1.28
readme-renderer==26.0
requests==2.22.0
requests-negotiate-sspi==0.5.2
scikit-learn==0.23.1
scipy==1.4.1
scrypt==0.8.15
setupcfg==2019.4.13
sha1==1.2
six==1.14.0
SQLAlchemy==1.3.11
sqlalchemy-pyodbc-mssql==0.1.0
tempora==3.0.0
threadpoolctl==2.1.0
urllib3==1.25.8
values==2019.4.13
visitor==0.1.3
wcwidth==0.2.4
webencodings==0.5.1
WebOb==1.8.5
Werkzeug==1.0.0
write==2019.4.13
wsgi-basic-auth==1.1.0
wsgi-fileserver==0.2.8
WTForms==2.2.1
xlrd==1.2.0
zc.lockfile==2.0
zipp==3.1.0
